Add ortho lights (still has serious issues that make them unusable on regular maps). First real attempt at lit water. Parsing rtlights is now aware of spotlights. Default cl_yieldcpu to 1, to save cpu for anyone who sets cl_maxfps lower. Added to menus. presets now include view angle clamping - maps made for quakespasm REQUIRE full pitch angles despite it otherwise being considered a cheat or glitchy (on servers that try to block the cheat). fix r_fullbrightSkins>=1 issue.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5274 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/cl_input.c
@@ -1765,9 +1765,9 @@ static void CL_SendUserinfoUpdate(void)
 		else if (cls.fteprotocolextensions2 & PEXT2_INFOBLOBS)
 		{	//only flood servers that actually support it.
 			if (final)
-				s = va("%ssetinfo \"%s\" \"%s\" %u", pl, enckey, encval, bloboffset);
+				s = va("%ssetinfo \"%s\" \"%s\" %u", pl, enckey, encval, (unsigned int)bloboffset);
 			else
-				s = va("%ssetinfo \"%s\" \"%s\" %u+", pl, enckey, encval, bloboffset);
+				s = va("%ssetinfo \"%s\" \"%s\" %u+", pl, enckey, encval, (unsigned int)bloboffset);
 		}
 		else
 		{	//server doesn't support it, just ignore the key
